Update optimizer tests.

+++ tests/optimizer/7/test.hs
@@ -1,3 +1,3 @@
 module Test where {
-f x = let {h x y = g y; g y = k y; k y = h 1 y} in g x;
+f x = let {h x y = g y; g y = k y; k y = h (1::Int) y} in g x;
 }
